added parent class contact
